requirements.txt
mcp
langchain
langchain-community
langchain-openai
chromadb
python-dotenv
tiktoken
pydantic